Add resume continuation test; currently fails
